Moving static maps to standalone library component
